requirements-gen.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
apache-airflow==2.10.5
apache-airflow-providers-common-compat==1.3.0
apache-airflow-providers-common-io==1.5.0
apache-airflow-providers-common-sql==1.21.0
apache-airflow-providers-fab==1.5.3
apache-airflow-providers-ftp==3.12.0
apache-airflow-providers-google==12.0.0
apache-airflow-providers-http==5.0.0
apache-airflow-providers-imap==3.8.0
apache-airflow-providers-smtp==1.9.0
apache-airflow-providers-sqlite==4.0.0
apispec==6.8.1
appdirs==1.4.4
argcomplete==3.5.3
asgiref==3.8.1
async-timeout==5.0.1
asyncssh==2.20.0
atpublic==5.1
attrs==25.1.0
autocorrect==2.6.1
babel==2.17.0
backoff==2.2.1
billiard==4.2.1
blinker==1.9.0
cachelib==0.9.0
cachetools==5.5.1
cattrs==24.1.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
colorama==0.4.6
colorlog==6.9.0
configobj==5.0.9
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==43.0.3
db-dtypes==1.4.1
decorator==5.1.1
Deprecated==1.2.18
deprecation==2.1.0
dictdiffer==0.9.0
dill==0.3.9
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
dpath==2.2.0
dulwich==0.22.7
dvc==3.59.1
dvc-data==3.16.9
dvc-gs==3.0.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
email_validator==2.2.0
entrypoints==0.4
eval_type_backport==0.2.2
exceptiongroup==1.2.2
filelock==3.17.0
Flask==2.2.5
Flask-AppBuilder==4.5.3
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
flatten-dict==0.4.2
flufl.lock==8.1.0
frozenlist==1.5.0
fsspec==2025.2.0
funcy==2.0
gcloud-aio-auth==5.3.2
gcloud-aio-bigquery==7.1.0
gcloud-aio-storage==9.3.0
gcsfs==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
google-ads==25.1.0
google-analytics-admin==0.23.4
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.81.0
google-cloud-alloydb==0.4.2
google-cloud-appengine-logging==1.6.0
google-cloud-audit-log==0.3.0
google-cloud-automl==2.16.0
google-cloud-batch==0.17.34
google-cloud-bigquery==3.29.0
google-cloud-bigquery-datatransfer==3.19.0
google-cloud-bigtable==2.28.1
google-cloud-build==3.30.0
google-cloud-compute==1.25.0
google-cloud-container==2.56.0
google-cloud-core==2.4.1
google-cloud-datacatalog==3.25.0
google-cloud-dataflow-client==0.8.16
google-cloud-dataform==0.5.15
google-cloud-dataplex==2.7.0
google-cloud-dataproc==5.17.0
google-cloud-dataproc-metastore==1.18.0
google-cloud-dlp==3.27.0
google-cloud-kms==3.3.0
google-cloud-language==2.17.0
google-cloud-logging==3.11.4
google-cloud-memcache==1.12.0
google-cloud-monitoring==2.27.0
google-cloud-orchestration-airflow==1.17.0
google-cloud-os-login==2.17.0
google-cloud-pubsub==2.28.0
google-cloud-redis==2.17.0
google-cloud-resource-manager==1.14.0
google-cloud-run==0.10.15
google-cloud-secret-manager==2.23.0
google-cloud-spanner==3.51.0
google-cloud-speech==2.31.0
google-cloud-storage==2.19.0
google-cloud-storage-transfer==1.16.0
google-cloud-tasks==2.19.0
google-cloud-texttospeech==2.25.0
google-cloud-translate==3.20.0
google-cloud-videointelligence==2.16.0
google-cloud-vision==3.10.0
google-cloud-workflows==1.17.0
google-crc32c==1.6.0
google-re2==1.1.20240702
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
gotrue==2.11.4
grandalf==0.8
greenlet==3.1.1
grpc-google-iam-v1==0.14.0
grpc-interceptor==0.15.4
grpcio==1.70.0
grpcio-gcp==0.2.2
grpcio-status==1.70.0
gto==1.7.2
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hydra-core==1.3.2
hyperframe==6.1.0
idna==3.10
immutabledict==4.2.1
importlib_metadata==8.5.0
inflection==0.5.1
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
json-merge-patch==0.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kneed==0.8.5
kombu==5.4.2
lazy-object-proxy==1.10.0
limits==4.0.1
linkify-it-py==2.0.3
lockfile==0.12.2
looker-sdk==25.2.0
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.6.0
multidict==6.1.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
ordered-set==4.1.0
orjson==3.10.15
packaging==24.2
pandas==2.1.4
pandas-gbq==0.27.0
pathspec==0.12.1
pendulum==3.0.0
platformdirs==4.3.6
pluggy==1.5.0
postgrest==0.19.3
prison==0.2.1
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydata-google-auth==1.9.1
pydot==3.0.4
pygit2==1.15.1
Pygments==2.19.1
pygtrie==2.5.0
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
realtime==2.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.0
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.1
scipy==1.13.1
scmrepo==3.3.10
semver==3.0.4
setproctitle==1.3.4
shapely==2.0.7
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==1.4.54
sqlalchemy-bigquery==1.12.1
SQLAlchemy-JSONField==1.0.2
sqlalchemy-spanner==1.8.0
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
sqltrie==0.11.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.5.0
time-machine==2.16.0
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
universal_pathlib==0.2.6
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
websockets==14.2
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.17.2
WTForms==3.2.1
yarl==1.18.3
zc.lockfile==3.0.post1
zipp==3.21.0
